ahora podemos agregar productos meduante al escaneo del codigo de barras, lo que si antes de implementarlo en las demas ventana como la venatana de alertas, tengo que modificar algunas ventanas, ya que se esta utilizandon un dispositivo android en el cual tiene muchas diferencias de tamaño con el emulador que se esta ocupando, si mas bien en el emulador se ve bien, pero en el telefono android se ve todo un poco movido, asi que primero se trabajara para adpatar la aplicacion para el dispositivo movil y luego pasaremos a agregar el apartado de escane en el resto de las ventanas

+++ app/src/main/java/com/example/proyectoprueba/controller/agregarAlertas.java
@@ -72,7 +72,7 @@ public class agregarAlertas extends AppCompatActivity {
             return;
         }
 
-        int codigoBarras;
+        long codigoBarras;
         int cantidad;
         int stockBodega;
         int stockGondola;
